% Auto-generated: do not edit by hand
\name{''DashReactEditableJsonTree}

\alias{''DashReactEditableJsonTree}

\title{DashReactEditableJsonTree component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
''DashReactEditableJsonTree(id=NULL, data=NULL, readOnly=NULL, updatedData=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{data}{Named list | unnamed list. JSON data}

\item{readOnly}{Logical. Whether the whole structure will be readonly}

\item{updatedData}{Named list | unnamed list. Updated JSON data}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

